优化curl下载

+++ app/src/main/assets/kr-script/Start.sh
@@ -15,13 +15,11 @@ Install_Curl() {
         if $Have_ROOT; then
             . $Core
             downloader "$ELF1_Path/curl3" "$Curl_URL/`getprop ro.product.cpu.abi`"
-            [[ ! -f "$ELF1_Path/curl3" ]] && downloader "$ELF1_Path/curl3" "$Curl_URL/armeabi-v7a"
             downloader ~/cacert.pem "$Curl_URL/cacert.pem"
-        else
-            wget -O $ELF1_Path/curl3 "$Curl_URL/`getprop ro.product.cpu.abi`"
-            [[ ! -f "$ELF1_Path/curl3" ]] && wget -O $ELF1_Path/curl3 "$Curl_URL/armeabi-v7a"
-            wget -O ~/cacert.pem "$Curl_URL/cacert.pem"
         fi
+        [[ ! -f "$ELF1_Path/curl3" ]] && wget -O $ELF1_Path/curl3 "$Curl_URL/`getprop ro.product.cpu.abi`"
+        [[ ! -f "$ELF1_Path/curl3" ]] && wget -O $ELF1_Path/curl3 "$Curl_URL/armeabi-v7a"
+        [[ ! -f ~/cacert.pem ]] && wget -O ~/cacert.pem "$Curl_URL/cacert.pem"
         chmod +x $ELF1_Path/curl3
 }
